requirements.txt
streamlit>=1.28.0
langchain>=0.0.267
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
qdrant-client>=1.6.0
uuid>=1.30
airtable-python-wrapper>=0.15.3
